requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
apscheduler
python-dotenv
pydantic
openai
